+++ glassure/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf8 -*-
-__version__ = '1.3.0'
+__version__ = '1.3.0.post13.dev0+f777488'