+++ glassure/gui/controller/glassure_controller.py
@@ -6,11 +6,10 @@ from PySide6.QtCore import Slot
 from qtpy import QtCore, QtWidgets
 import pyqtgraph as pg
 
-from ..widgets.glassure import GlassureWidget
+from ..widgets.glassure_widget import GlassureWidget
 from ..widgets.custom.file_dialogs import open_file_dialog, save_file_dialog
 
-from ..model.glassure import GlassureModel
-from ...core.scattering_factors import set_source as set_scattering_data_source
+from ..model.glassure_model import GlassureModel
 from ...core.scattering_factors import get_available_elements
 
 from .configuration import ConfigurationController
@@ -151,18 +150,22 @@ class GlassureController(object):
         cur_ind = self.main_widget.left_control_widget.composition_widget.composition_tw.currentRow()
         self.main_widget.left_control_widget.composition_widget.delete_element(cur_ind)
 
-    def data_source_changed(self, text):
-        set_scattering_data_source(text)
+    def data_source_changed(self, source: str):
+        self.validate_composition(source)
+        self.update_model()
+
+    def validate_composition(self, source):
         composition = self.main_widget.get_composition()
         for element in list(composition.keys()):
-            if element not in get_available_elements():
+            if element not in get_available_elements(source):
                 del composition[element]
         self.main_widget.left_control_widget.composition_widget.set_composition(composition)
-        self.update_model()
 
     @Slot()
     def update_model(self):
         composition = self.main_widget.get_composition()
+        sf_source = self.main_widget.get_sf_source()
+
         density = self.main_widget.left_control_widget.composition_widget.get_density()
 
         q_min, q_max, r_min, r_max = self.main_widget.get_parameter()
@@ -171,9 +174,11 @@ class GlassureController(object):
         extrapolation_method = self.main_widget.get_extrapolation_method()
         extrapolation_parameters = self.main_widget.get_extrapolation_parameters()
 
-        optimize_active, r_cutoff, optimize_iterations, optimize_attenuation = self.main_widget.get_optimization_parameter()
+        optimize_active, r_cutoff, optimize_iterations, optimize_attenuation = \
+            self.main_widget.get_optimization_parameter()
 
-        self.model.update_parameter(composition, density,
+        self.model.update_parameter(sf_source,
+                                    composition, density,
                                     q_min, q_max,
                                     r_min, r_max,
                                     use_modification_fcn,