+++ glassure/gui/widgets/glassure_widget.py
@@ -104,6 +104,9 @@ class GlassureWidget(QtWidgets.QWidget):
     def create_function_shortcuts(self):
         self.set_composition = self.left_control_widget.composition_widget.set_composition
         self.get_composition = self.left_control_widget.composition_widget.get_composition
+        self.get_sf_source = self.left_control_widget.composition_widget.get_sf_source
+        self.set_sf_source = self.left_control_widget.composition_widget.set_sf_source
+
         self.get_density = self.left_control_widget.composition_widget.get_density
 
         self.get_parameter = self.left_control_widget.options_widget.get_parameter