+++ glassure/run.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import
 import sys
 
 from glassure import __version__
-from glassure.gui.controller.glassure import GlassureController
+from glassure.gui.controller.glassure_controller import GlassureController
 from qtpy import QtWidgets
 
 
@@ -14,7 +14,7 @@ def my_exception_hook(exctype, value, traceback):
     print(exctype, value, traceback)
 
 
-sys.excepthook = my_exception_hook
+# sys.excepthook = my_exception_hook
 
 
 def main():